refactor API files after feedback
