+++ src/js/commonFunctions/api/auth/index.mjs
@@ -1,0 +1,2 @@
+export * from "./login.mjs";
+export * from "./signup.mjs";
