+++ src/js/commonFunctions/api/profiles/index.mjs
@@ -1,0 +1,2 @@
+export * from "./read.mjs";
+export * from "./update.mjs";
